requirements.txt
Flask==1.1.1
Flask-Login==0.4.1
flask-redis==0.4.0
Flask-Sessions==0.1.5
mysqlclient==2.1.0
lxml==4.4.2
redis==3.3.11
requests==2.22.0
celery==4.4.0
